Rename the pynput library to magicy in order to avoid Windows Real-time protection
