+++ README.md
@@ -1,1 +1,3 @@
 # Migaku Anki
+
+Note: We renamed the pynput library to magicy in order to avoid Windows Real-time protection
