+++ src/lib/shared/magicy/__init__.py
