+++ src/lib/shared/magicy/_info.py
