+++ src/lib/shared/magicy/_util/darwin.py
