+++ src/lib/shared/magicy/_util/darwin_vks.py
