+++ src/lib/shared/magicy/_util/uinput.py
