+++ src/lib/shared/magicy/_util/win32.py
