+++ src/lib/shared/magicy/_util/win32_vks.py
