+++ src/lib/shared/magicy/keyboard/__init__.py
@@ -25,7 +25,7 @@ See the documentation for more information.
 
 import itertools
 
-from pynput._util import backend, Events
+from magicy._util import backend, Events
 
 
 backend = backend(__name__)
