+++ src/lib/shared/magicy/keyboard/_base.py
@@ -31,8 +31,8 @@ import unicodedata
 
 import six
 
-from pynput._util import AbstractListener, prefix
-from pynput import _logger
+from magicy._util import AbstractListener, prefix
+from magicy import _logger
 
 
 class KeyCode(object):
@@ -733,7 +733,7 @@ class Listener(AbstractListener):
         :return: a key
         :rtype: Key or KeyCode
         """
-        from pynput.keyboard import Key, KeyCode, _NORMAL_MODIFIERS
+        from magicy.keyboard import Key, KeyCode, _NORMAL_MODIFIERS
 
         if isinstance(key, KeyCode) and key.char is not None:
             return KeyCode.from_char(key.char.lower())
