+++ src/lib/shared/magicy/keyboard/_darwin.py
@@ -28,12 +28,12 @@ import enum
 
 import Quartz
 
-from pynput._util.darwin import (
+from magicy._util.darwin import (
     get_unicode_to_keycode_map,
     keycode_context,
     ListenerMixin,
 )
-from pynput._util.darwin_vks import SYMBOLS
+from magicy._util.darwin_vks import SYMBOLS
 from . import _base
