+++ src/lib/shared/magicy/mouse/__init__.py
@@ -23,7 +23,7 @@ See the documentation for more information.
 # pylint: disable=C0103
 # Button, Controller and Listener are not constants
 
-from pynput._util import backend, Events
+from magicy._util import backend, Events
 
 
 backend = backend(__name__)
